完善消息列表，医生咨询，健康，医疗图表等逻辑

+++ apilib/src/main/java/com/guider/health/apilib/bean/BloodSugarListBean.kt
@@ -3,7 +3,7 @@ package com.guider.health.apilib.bean
 data class BloodSugarListBean(
     val accountId: Int,
     val bloodSpeed: Int,
-    val bs: Float,
+    val bs: Double,
     val bsTime: String,
     val createTime: String,
     val deviceCode: String,
